requirements.txt
ultralytics
opencv-python
numpy
pillow
piexif
torch
torchvision
